refactor(Board): 메서드가 한가지 일을 할 수 있도록 분리
